Added plugin downloader script with spiget
This script doesn't support premium plugins, at least for now. But it can download any amount of free spigot plugins just by adding links on the "mcsplugins" file. The purpose of this script will be to create server configuration packs in a legal way, as jar files won't need to be shared. I personally use it to upload my server configurations to GitHub without uploading the jar files.